# schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Northwind MCP API",
    "version": "1.0.0",
    "description": "Comprehensive OpenAPI specification for the Northwind MCP Server, exposing business analytics, summaries, and entity data for Copilot Studio integration."
  },
  "servers": [
    {
      "url": "https://nr-northwind-mcp-fhauhhduczeea6ft.canadacentral-01.azurewebsites.net",
      "description": "Northwind MCP Production API"
    }
  ],
...
  "paths": {
...
    "/customers": {
      "get": {
        "tags": ["Core", "Financial"],
        "operationId": "getCustomers",
        "summary": "Get Customers",
        "description": "Retrieve all customers, including company names, cities, and contact details.",
        "parameters": [
          { "name": "country", "in": "query", "schema": { "type": "string" }, "description": "Filter by country" },
          { "name": "city", "in": "query", "schema": { "type": "string" }, "description": "Filter by city" },
          { "name": "customer_id", "in": "query", "schema": { "type": "string" }, "description": "Filter by customer ID" },
          { "name": "company_name", "in": "query", "schema": { "type": "string" }, "description": "Filter by company name" },
          { "name": "contact_name", "in": "query", "schema": { "type": "string" }, "description": "Filter by contact name" },
          { "name": "contact_title", "in": "query", "schema": { "type": "string" }, "description": "Filter by contact title" },
          { "name": "address", "in": "query", "schema": { "type": "string" }, "description": "Filter by address" },
          { "name": "region", "in": "query", "schema": { "type": "string" }, "description": "Filter by region" },
          { "name": "postal_code", "in": "query", "schema": { "type": "string" }, "description": "Filter by postal code" },
          { "name": "phone", "in": "query", "schema": { "type": "string" }, "description": "Filter by phone" },
          { "name": "fax", "in": "query", "schema": { "type": "string" }, "description": "Filter by fax" }
        ],
        "responses": {
          "200": {
            "description": "List of customers.",
            "content": {
              "application/json": {
                "example": {
                  "filters": { "country": "Germany", "city": "Berlin" },
                  "count": 1,
                  "data": [
                    { "customer_id": "ALFKI", "company_name": "Alfreds Futterkiste", "city": "Berlin", "country": "Germany" }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}